+ module-init-tools patch by Sergey Vlasov:
The insmod program from module-init-tools is very simple and does not
support any options, includind the "-o NAME" option to rename the
loaded module - modprobe should be used instead.  However, in the old
modutils package the "-o NAME" option was supported only by insmod and
not modprobe.  This patch adds the modprobe_with_rename function which
detects the variant of modprobe in use and invokes either insmod or
modprobe, and replaces all direct $INSMOD calls with the "-o NAME"
option with calls to modprobe_with_rename.



git-svn-id: svn://svn.etcnet.org/trunk/etcnet@996 4788b5ed-9404-0410-8295-8f3c5ce22fc0